Move job interface to JobClient
