+++ frontend/src/job/JobClient.ts
@@ -1,6 +1,5 @@
 import { JobParameters } from './JobParameters';
 import { sendRequest } from '@/util/HTTPClient';
-import { Job } from './Job';
 import { backendUri } from '@/config';
 
 export async function startJob(
@@ -10,10 +9,26 @@ export async function startJob(
     return sendRequest('post', `${backendUri}/job/${jobType}`, params, false);
 }
 
-export async function getJobList(get_all:boolean): Promise<Job[]> {
-    return sendRequest('get', `${backendUri}/job/jobs`, {show_all_jobs:get_all}, false);
+export async function getJobList(getAllJobs: boolean): Promise<Job[]> {
+    return sendRequest('get', `${backendUri}/job/jobs`, { show_all_jobs: getAllJobs }, false);
 }
 
 export async function getJobDetails(jobID: string): Promise<Job> {
     return sendRequest('get', `${backendUri}/job/${jobID}`, {}, false);
+}
+
+export interface Job {
+    children: object[];
+    created: string;
+    duration:string;
+    errors: object[]; // TODO proper error interface
+    job_id: string; // eslint-disable-line camelcase
+    job_type: string; // eslint-disable-line camelcase
+    name: string;
+    parameters: object;
+    parent_job_id:string; // eslint-disable-line camelcase
+    started:string;
+    state: string;
+    updated: string;
+    user: string;
 }
