+++ frontend/src/job/JobListCompact.vue
@@ -31,8 +31,7 @@
 
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
-import { getJobList } from './JobClient';
-import { Job } from './Job';
+import { getJobList, Job } from './JobClient';
 import { showError } from '@/util/Notifier.ts';
 
 @Component
